feat: add Fuga Cloud
