+++ config/nukkit/server-fuga.yaml
@@ -1,0 +1,56 @@
+apiVersion: minectl.ediri.io/v1alpha1
+kind: MinecraftServer
+metadata:
+  name: minecraft-server
+spec:
+  server:
+    cloud: fuga
+    region: ams2
+    size: s3.large
+    ssh:
+      port: 22
+      keyfolder: "/Users/dirien/Tools/repos/stackit-minecraft/minecraft/ssh/minecraft"
+      fail2ban:
+        bantime: 1000
+        maxretry: 3
+    port: 19132
+  minecraft:
+    java:
+      openjdk: 8
+      xmx: 2G
+      xms: 2G
+      rcon:
+        password: test
+        port: 25575
+        enabled: true
+        broadcast: true
+    edition: nukkit
+    version: 1.0-SNAPSHOT
+    eula: true
+    properties: |
+      motd=Fuga Cloud Minecraft --- Nukkitx edition
+      sub-motd=https://nukkitx.com
+      server-ip=0.0.0.0
+      view-distance=10
+      white-list=off
+      achievements=on
+      announce-player-achievements=on
+      spawn-protection=16
+      max-players=20
+      allow-flight=off
+      spawn-animals=on
+      spawn-mobs=on
+      gamemode=0
+      force-gamemode=off
+      hardcore=off
+      pvp=on
+      difficulty=1
+      generator-settings=
+      level-name=world
+      level-seed=
+      level-type=DEFAULT
+      allow-nether=on
+      enable-query=on
+      auto-save=on
+      force-resources=off
+      xbox-auth=on
